Aspen Capital/requirements.txt
flask
mysql-connector-python
tabulate
